# specs/002-admin-panel-telegram-auth/contracts/auth-admin-api.openapi.yaml
openapi: 3.1.0
info:
  title: BasmatLibya — Auth + Admin API (feature 002)
  version: 0.2.0
  description: |
    Additive REST surface for feature `002-admin-panel-telegram-auth`. This document
    describes ONLY the new endpoints introduced by this feature: Telegram-based
    authentication, signed-in user history, and the owner-only admin panel.

    The v1 surface from `001-digital-footprint-analyzer` (`/api/lookups`,
    `/api/lookups/:id`, `/api/lookups/:id/rerun`, `/api/healthz`) remains
    unchanged. All schemas here are generated from `zod` schemas in
    `packages/shared`; treat this file as a snapshot — the runtime is the source
    of truth.

    Authentication: a server-issued opaque session cookie `bsl_session`
    (HttpOnly; Secure; SameSite=Lax). State-changing endpoints additionally
    require an `X-CSRF` header equal to the session's `csrfToken` (returned
    by `/api/auth/me`). The Telegram callback `/api/auth/telegram` is exempt
    from `X-CSRF` because it carries its own HMAC verification.

    Authorization roles:
      - anonymous: no cookie or revoked cookie.
      - user:     valid cookie, role='user', status='active'.
      - owner:    valid cookie, role='owner', status='active'.

    All user-facing strings on `4xx`/`5xx` errors are Arabic and pre-translated
    server-side; clients render them verbatim.
servers:
  - url: https://basmatlibya.example.com
    description: Production (Render)
  - url: http://localhost:5173
...
paths:
...
  /api/admin/users:
    get:
      tags: [admin-users]
      summary: List users with filters and pagination
      operationId: listUsers
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, suspended] }
        - in: query
          name: role
          schema: { type: string, enum: [owner, user] }
        - in: query
          name: sort
          schema: { type: string, enum: [joined_desc, last_seen_desc], default: last_seen_desc }
        - in: query
          name: cursor
          schema: { type: string, nullable: true }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Users page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUsersPage' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/admin/users/{id}:
    get:
      tags: [admin-users]
      summary: Read a single user with their lookup history (FR-020)
      operationId: getAdminUser
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: User detail with paginated history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUserDetail' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [admin-users]
      summary: Remove a user
      description: |
        Cascades sessions and `user_lookup_associations`; sets
        `lookups.owner_user_id` to NULL (share links survive).
        Blocked when the target is the last/only owner (FR-010) or when
        the target has any `ai_model_entries` they created — those must
        be deleted or reassigned first.
      operationId: deleteAdminUser
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Removed.
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Cannot remove (last owner, or has dependent AI model entries).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/admin/users/{id}/suspend:
    post:
      tags: [admin-users]
      summary: Suspend a user (revokes every active session of the user)
      description: |
        Transactionally sets `users.status='suspended'`, sets `revoked_at`
        on every active session of that user, appends an audit row, and
        emits `session.invalidated` to the user's socket room.
      operationId: suspendUser
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Suspended.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUserSummary' }
        '403':
          description: Caller is not the owner, or attempting to self-suspend.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/admin/users/{id}/unsuspend:
    post:
      tags: [admin-users]
      summary: Lift a user's suspension
      operationId: unsuspendUser
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Reinstated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUserSummary' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  parameters:
    CsrfHeader:
      in: header
      name: X-CSRF
      required: true
      schema: { type: string }
      description: Per-session double-submit token from `/api/auth/me` (R-04).
  schemas:
...
    LookupHistoryItem:
      type: object
      required: [lookupId, identifierType, identifierPreview, status, createdAt, expiresAt]
      properties:
        lookupId:          { type: string, format: uuid }
        identifierType:    { type: string, enum: [name, username, email, phone] }
        identifierPreview: { type: string, description: "Truncated for display." }
        status:            { type: string, enum: [in_progress, completed, failed, cancelled, expired] }
        createdAt:         { type: string, format: date-time }
        expiresAt:         { type: string, format: date-time }
...
    AdminUserSummary:
      type: object
      required: [id, telegramId, displayName, role, status, joinedAt, lastSeenAt]
      properties:
        id:          { type: string, format: uuid }
        telegramId:  { type: integer, format: int64 }
        displayName: { type: string }
        username:    { type: string, nullable: true }
        avatarUrl:   { type: string, format: uri, nullable: true }
        role:        { type: string, enum: [owner, user] }
        status:      { type: string, enum: [active, suspended] }
        joinedAt:    { type: string, format: date-time }
        lastSeenAt:  { type: string, format: date-time }
        suspendedAt: { type: string, format: date-time, nullable: true }
    AdminUserDetail:
      allOf:
        - $ref: '#/components/schemas/AdminUserSummary'
        - type: object
          required: [history, activeSessions]
          properties:
            history:
              type: array
              items: { $ref: '#/components/schemas/LookupHistoryItem' }
            activeSessions: { type: integer, minimum: 0 }
    AdminUsersPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AdminUserSummary' } }
        nextCursor: { type: string, nullable: true }
...
    ErrorResponse:
      type: object
      required: [code, messageAr]
      properties:
        code:
          type: string
          description: Stable machine code.
          enum:
            - not_authenticated
            - unauthorized
            - csrf_required
            - hmac_invalid
            - auth_date_too_old
            - bot_unavailable
            - suspended_user
            - validation_failed
            - last_owner_protected
            - dependent_entries
            - active_model_protected
            - rate_limited
            - not_found
        messageAr:
          type: string
          description: Pre-translated Arabic message; the UI renders verbatim.
        details:
          type: object
          nullable: true
          description: Optional structured details (e.g. per-field validation errors).